Allow float distances

+++ src/org/vika/routing/network/Channel.java
@@ -7,7 +7,7 @@ import java.util.Scanner;
  */
 public class Channel {
     public final int id;
-    public int time;
+    public float time;
 
     public Channel(final int id) {
         this.id = id;
@@ -15,7 +15,7 @@ public class Channel {
 
     public static Channel parse(final int id, final Scanner scanner) {
         final Channel channel = new Channel(id);
-        channel.time = scanner.nextInt();
+        channel.time = scanner.nextFloat();
         return channel;
     }
 }
